By moving the trigger for the shutdown of the `EmbeddedEventStore` from using an annotation, to implementing `LifeCycle` it's no longer `javax` or `jakarta` specific. This makes sure it always works.
